Refactor formatting.
